Updates in this commit
- Fully set up main menu functionality
- Fully set up settings menu functionalities
- Imported UI assets

+++ Assets/_Scripts/mad/Dummies/DummySettingsTester.cs
@@ -5,9 +5,9 @@ using WGRF.Core;
 public class DummySettingsTester : MonoBehaviour
 {
     public bool gore;
-    public float master;
-    public float ost;
-    public float sfx;
+    public int master;
+    public int ost;
+    public int sfx;
 
     // Update is called once per frame
     void Update()
